# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WeSwing",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/M14/Projecte_Final/grup2/server/public/index.php"
        }
    ],
    "paths": {
...
        "/api/usuaris": {
            "get": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Veure tots els usuaris.",
                "operationId": "App\\Http\\Controllers\\UsuariController::getUsuaris",
                "responses": {
                    "200": {
                        "description": "Retorna tots els usuaris."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
        "/api/usuari/{correo}": {
            "get": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Veure l'usuari desitjat.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\UsuariController::getUsuariCorreo",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "usuari per correu",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuari amb email seleccionat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/usuari/{id}": {
            "post": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Actualitza l'usuari ja existent",
                "description": "",
                "operationId": "updateUsuari",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id usuari a modificar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "entitat_id",
                        "in": "query",
                        "description": "Id de l'entitat",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "persona_id",
                        "in": "query",
                        "description": "Id de la persona",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "Nickname",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contrasenya",
                        "in": "query",
                        "description": "Contrasenya",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dataNaixement",
                        "in": "query",
                        "description": "Data de Naixement",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcio",
                        "in": "query",
                        "description": "Descripcio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "imagen",
                        "in": "query",
                        "description": "Imatge",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idioma",
                        "in": "query",
                        "description": "Idioma",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idioma",
                        "in": "query",
                        "description": "Idioma",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "genere",
                        "in": "query",
                        "description": "Genere",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pais",
                        "in": "query",
                        "description": "Pais",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vacunaCovid",
                        "in": "query",
                        "description": "Vacuna Covid-19",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuari modificat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Eliminar un usuari.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\UsuariController::eliminarUsuari",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Usuari a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuari eliminat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/usuari": {
            "post": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Inserir un nou usuari.",
                "operationId": "App\\Http\\Controllers\\UsuariController::crearUsuari",
                "parameters": [
                    {
                        "name": "enitat_id",
                        "in": "query",
                        "description": "Id de l'entitat",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "persona_id",
                        "in": "query",
                        "description": "Id de la persona",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "Nickname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contrasenya",
                        "in": "query",
                        "description": "Contrasenya",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dataNaixement",
                        "in": "query",
                        "description": "Data de Naixement",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcio",
                        "in": "query",
                        "description": "Descripcio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "imagen",
                        "in": "query",
                        "description": "Imatge",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idioma",
                        "in": "query",
                        "description": "Idioma",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idioma",
                        "in": "query",
                        "description": "Idioma",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "genere",
                        "in": "query",
                        "description": "Genere",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pais",
                        "in": "query",
                        "description": "Pais",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vacunaCovid",
                        "in": "query",
                        "description": "Vacuna Covid-19",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna l'usuari que hem inserit."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
        "/api/usuariId/{id}": {
            "get": {
                "tags": [
                    "Usuaris"
                ],
                "summary": "Veure l'usuari desitjat.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\UsuariController::getUsuari",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Usuari per id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuari amb id seleccionat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}